update and cleanup library linking in cmakelists. fix type mismatches that occur in 64bit due to differences in unsigned and size_t
--HG--
branch : cmake-64bit

+++ scone/cs/FeedForwardController.cpp
@@ -24,7 +24,7 @@ namespace scone
 		Controller( props, par, model, target_area )
 		{
 			DECLARE_AND_INIT( props, bool, use_symmetric_actuators, true );
-			INIT_PROPERTY( props, number_of_modes, 0u );
+			INIT_PROPERTY( props, number_of_modes, size_t( 0 ) );
 
 			// setup actuator info
 			for ( size_t idx = 0; idx < model.GetMuscles().size(); ++idx )
